ENH: make valview list-like instead of set-like.
Fixes the py2/3 compat issues with operators on view objects.
Uses more advanced templates to only populate the operators for the
given type. For example, valview does not suppport subtract because it
is list like. Instead of having a function in the slot that raised this
now properly returns NotImplemented.